Fast API project/backend-secure-video-streaming/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
cryptography==41.0.1
psycopg2-binary==2.9.6
kafka-python==2.0.2
python-dotenv==1.0.0